Updated Index.html and about.html
Change the size, and coloring of the header title, and centered the text. Reduced the size of the header bar. Fixed contact us navbar link to be working, and removed the "about" navbar tabbed which is no longer needed. Changed the color of the buttons on the homepage to better fit with our co.or scheme.